Move app/mod.rs free helpers into src/app/util.rs
Pulls SCROLL_LINES, scroll_action, read_rss_kb (all four cfg-gated
platform overloads), parse_lsp_position, same_file, completion_kind_label,
extract_hover_text, parse_locations and extract_word_at out of app/mod.rs
and into a sibling util.rs. They're pub(super) so only the app module
sees them. A use line at the top of mod.rs keeps the 19 existing call
sites bare. No behaviour change.
